requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.17.1
pillow==10.2.0
numpy>=1.26.0
python-dotenv==0.19.0